+++ base/Module.py
@@ -12,10 +12,12 @@ class Module:
     base.VERSION.added("1.1.1", "base.Module class for describing Landscape Model modules")
     base.VERSION.changed("1.3.5", "base.Module refactored")
     base.VERSION.added("1.4.1", "Changelog in base.Module")
+    base.VERSION.added("1.5.0", "`base.Module.doc_file` ")
 
-    def __init__(self, name, version):
+    def __init__(self, name, version, doc_file=None):
         self._name = name
         self._version = version
+        self._doc_file = doc_file
         return
 
     @property
@@ -31,3 +33,10 @@ class Module:
         Gets the version of the module as specified by the module author.
         """
         return self._version
+
+    @property
+    def doc_file(self):
+        """
+        Gets the path of the additional documentation file, if available.
+        """
+        return self._doc_file